Add startServer function
